Разделил EditWordDefinitionsRepository на два репозитория
Теперь есть 2 отдельных репозитория LookupWordDefinitionsRepository, загружающий поределения по написанию, и EditWordDefinitionsRepository, используемый для редактирования определений слов
