+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/application/AppComponent.kt
@@ -10,6 +10,7 @@ import com.designdrivendevelopment.kotelok.repositoryImplementations.getStatisti
 import com.designdrivendevelopment.kotelok.repositoryImplementations.learnableDefinitionsRepository.CardsLearnableDefinitionsRepository
 import com.designdrivendevelopment.kotelok.repositoryImplementations.learnableDefinitionsRepository.PairsLearnableDefinitionsRepository
 import com.designdrivendevelopment.kotelok.repositoryImplementations.learnableDefinitionsRepository.WriterLearnableDefinitionsRepository
+import com.designdrivendevelopment.kotelok.repositoryImplementations.lookupWordDefinitionRepository.LookupWordDefRepositoryImpl
 import com.designdrivendevelopment.kotelok.screens.bottomNavigation.BottomNavigator
 import com.designdrivendevelopment.kotelok.screens.bottomNavigation.Tab
 import com.designdrivendevelopment.kotelok.screens.dictionaries.dictionariesScreen.DictionariesFragment
@@ -52,9 +53,14 @@ class AppComponent(applicationContext: Context) {
         PairsLearnableDefinitionsRepository(db.pairsLearnableDefDao)
     }
     val statisticsRepository by lazy { GetStatisticsRepositoryImpl(db.statisticsDao) }
+    val lookupWordDefRepository by lazy {
+        LookupWordDefRepositoryImpl(
+            yandexDictApiService = yandexDictionaryApiService,
+            wordDefinitionsDao = db.wordDefinitionsDao
+        )
+    }
     val editWordDefinitionsRepository by lazy {
         EditWordDefRepositoryImpl(
-            yandexDictApiService = yandexDictionaryApiService,
             wordDefinitionsDao = db.wordDefinitionsDao,
             dictWordDefCrossRefDao = db.dictionaryWordDefCrossRefDao,
             translationsDao = db.translationsDao,
