+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/repositoryImplementations/editWordDefnititionsRepository/EditWordDefRepositoryImpl.kt
@@ -11,64 +11,16 @@ import com.designdrivendevelopment.kotelok.persistence.roomEntities.DictionaryWo
 import com.designdrivendevelopment.kotelok.persistence.roomEntities.ExampleEntity
 import com.designdrivendevelopment.kotelok.persistence.roomEntities.SynonymEntity
 import com.designdrivendevelopment.kotelok.persistence.roomEntities.TranslationEntity
-import com.designdrivendevelopment.kotelok.repositoryImplementations.toWordDefinition
 import com.designdrivendevelopment.kotelok.repositoryImplementations.toWordDefinitionEntity
 import com.designdrivendevelopment.kotelok.screens.dictionaries.EditWordDefinitionsRepository
-import com.designdrivendevelopment.kotelok.yandexDictApi.YandexDictionaryApiService
-import com.designdrivendevelopment.kotelok.yandexDictApi.responses.YandexDictionaryResponse
-import java.net.UnknownHostException
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import retrofit2.HttpException
 
 class EditWordDefRepositoryImpl(
-    private val yandexDictApiService: YandexDictionaryApiService,
     private val wordDefinitionsDao: WordDefinitionsDao,
     private val dictWordDefCrossRefDao: DictionaryWordDefCrossRefDao,
     private val translationsDao: TranslationsDao,
     private val synonymsDao: SynonymsDao,
     private val examplesDao: ExamplesDao
 ) : EditWordDefinitionsRepository {
-    override fun loadDefinitionsByWriting(
-        writing: String
-    ): Flow<DefinitionsRequestResult> = flow {
-        try {
-            val response: YandexDictionaryResponse = yandexDictApiService.lookupWord(writing)
-            val definitionsList: List<WordDefinition> = response
-                .definitions.flatMap { definitionResponse ->
-                    definitionResponse.translations.map { translationResponse ->
-                        translationResponse.toWordDefinition(
-                            definitionResponse.writing,
-                            definitionResponse.transcription
-                        )
-                    }
-                }
-            emit(DefinitionsRequestResult.Success(definitionsList))
-        } catch (e: UnknownHostException) {
-            emit(DefinitionsRequestResult.Failure.Error("UnknownHostException"))
-        } catch (e: HttpException) {
-            emit(
-                DefinitionsRequestResult.Failure.HttpError(
-                    code = e.code(),
-                    message = e.message()
-                )
-            )
-        }
-    }.flowOn(Dispatchers.IO)
-
-    override fun getSavedDefinitionsByWriting(
-        writing: String
-    ): Flow<List<WordDefinition>> {
-        return wordDefinitionsDao.getFlowOfDefinitionsByWriting(writing).map { definitionsList ->
-            definitionsList.map { wordDefinitionQueryResult ->
-                wordDefinitionQueryResult.toWordDefinition()
-            }
-        }
-    }
-
     override suspend fun addWordDefinition(wordDefinition: WordDefinition) {
         if (wordDefinition.id == NEW_WORD_ID) {
             addNewWordDefinition(wordDefinition)
