+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/screens/dictionaries/dictionaryDetailsScreen/DictDetailsViewModelFactory.kt
@@ -2,7 +2,6 @@ package com.designdrivendevelopment.kotelok.screens.dictionaries.dictionaryDetai
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.designdrivendevelopment.kotelok.screens.dictionaries.DictionaryWordDefinitionsRepository
 
 class DictDetailsViewModelFactory(
     private val dictionaryId: Long,
