+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/screens/dictionaries/lookupWordDefinitionsScreen/LookupWordDefinitionsRepository.kt
@@ -1,0 +1,11 @@
+package com.designdrivendevelopment.kotelok.screens.dictionaries.lookupWordDefinitionsScreen
+
+import com.designdrivendevelopment.kotelok.entities.WordDefinition
+import com.designdrivendevelopment.kotelok.repositoryImplementations.lookupWordDefinitionRepository.DefinitionsRequestResult
+import kotlinx.coroutines.flow.Flow
+
+interface LookupWordDefinitionsRepository {
+    fun loadDefinitionsByWriting(writing: String): Flow<DefinitionsRequestResult>
+
+    fun getSavedDefinitionsByWriting(writing: String): Flow<List<WordDefinition>>
+}
